+++ Application/Words/Commands/Update/Update.Request.Validator.cs
@@ -23,6 +23,6 @@ public class UpdateRequestValidator : AbstractValidator<UpdateRequest>
     {
         using var dbContext = await _dbContextFactory.CreateDbContextAsync(cancellationToken);
         return !(await dbContext.Words
-            .AnyAsync(x => x.Value == value && x.Type == request.Type, cancellationToken));
+            .AnyAsync(x => x.Value == value && x.Type == (int)request.Type, cancellationToken));
     }
 }
